# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Procurement Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /eligibility/company-profile:
    get:
      operationId: getCompanyProfile
      tags: [eligibility]
      summary: Get company profile for eligibility matching
      responses:
        "200":
          description: Company profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyProfile"

    put:
      operationId: updateCompanyProfile
      tags: [eligibility]
      summary: Update company profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyProfile"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyProfile"
  /eligibility/check/{tenderId}:
    get:
      operationId: checkEligibility
      tags: [eligibility]
      summary: Check eligibility for a specific tender
      parameters:
        - name: tenderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Eligibility result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EligibilityResult"

  # ── BID WORKFLOW ──────────────────────────────────────────────────────────────
...
  /boq:
    get:
      operationId: listBoq
      tags: [boq]
      summary: List BOQ entries
      parameters:
        - name: bidId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: BOQ list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BoqItem"

    post:
      operationId: createBoqItem
      tags: [boq]
      summary: Add BOQ line item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBoqItemRequest"
      responses:
        "201":
          description: BOQ item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoqItem"
  /boq/{id}:
    put:
      operationId: updateBoqItem
      tags: [boq]
      summary: Update BOQ line item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBoqItemRequest"
      responses:
        "200":
          description: Updated BOQ item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoqItem"

    delete:
      operationId: deleteBoqItem
      tags: [boq]
      summary: Delete BOQ item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── DOCUMENTS ─────────────────────────────────────────────────────────────────
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents in vault
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Documents list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"

    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentRequest"
      responses:
        "201":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── KNOWLEDGE BASE ────────────────────────────────────────────────────────────
  /knowledge:
    get:
      operationId: listKnowledgeItems
      tags: [knowledge]
      summary: List knowledge base items
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Knowledge items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeItem"

    post:
      operationId: createKnowledgeItem
      tags: [knowledge]
      summary: Add knowledge base item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKnowledgeItemRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeItem"
  /knowledge/{id}:
    put:
      operationId: updateKnowledgeItem
      tags: [knowledge]
      summary: Update knowledge item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKnowledgeItemRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeItem"

    delete:
      operationId: deleteKnowledgeItem
      tags: [knowledge]
      summary: Delete knowledge item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── ANALYTICS ─────────────────────────────────────────────────────────────────
...
  /clarifications:
    get:
      operationId: listClarifications
      tags: [clarifications]
      summary: List clarification queries
      parameters:
        - name: bidId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, submitted, replied]
      responses:
        "200":
          description: Clarifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Clarification"

    post:
      operationId: createClarification
      tags: [clarifications]
      summary: Create a clarification question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClarificationRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clarification"
  /clarifications/{id}:
    put:
      operationId: updateClarification
      tags: [clarifications]
      summary: Update clarification status/reply
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClarificationRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clarification"

  # ── VENDORS ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    CompanyProfile:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        companyName:
          type: string
        gstNumber:
          type: string
        panNumber:
          type: string
        annualTurnover:
          type: number
        yearsOfExperience:
          type: integer
        certifications:
          type: array
          items:
            type: string
        oemAuthorizations:
          type: array
          items:
            type: string
        pastProjects:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              value:
                type: number
              authority:
                type: string
              year:
                type: integer
        categories:
          type: array
          items:
            type: string
        states:
          type: array
          items:
            type: string
      required: [companyName]
    EligibilityResult:
      type: object
      properties:
        tenderId:
          type: integer
        score:
          type: number
        eligible:
          type: boolean
        criteria:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              required:
                type: string
              present:
                type: string
              met:
                type: boolean
        gaps:
          type: array
          items:
            type: string
        suggestions:
          type: array
          items:
            type: string
      required: [tenderId, score, eligible]
...
    BoqItem:
      type: object
      properties:
        id:
          type: integer
        bidId:
          type: integer
        description:
          type: string
        unit:
          type: string
        quantity:
          type: number
        unitCost:
          type: number
        margin:
          type: number
        gst:
          type: number
        totalCost:
          type: number
        vendorQuote:
          type: number
        winProbability:
          type: number
      required: [id, bidId, description, quantity, unitCost]
    CreateBoqItemRequest:
      type: object
      properties:
        bidId:
          type: integer
        description:
          type: string
        unit:
          type: string
        quantity:
          type: number
        unitCost:
          type: number
        margin:
          type: number
        gst:
          type: number
        vendorQuote:
          type: number
      required: [bidId, description, quantity, unitCost]
    Document:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [gst, pan, iso, oem, invoice, work_order, certificate, other]
        fileUrl:
          type: string
        fileName:
          type: string
        expiryDate:
          type: string
          format: date
        isExpired:
          type: boolean
        isExpiringSoon:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, category, fileName]
    CreateDocumentRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [gst, pan, iso, oem, invoice, work_order, certificate, other]
        fileName:
          type: string
        fileUrl:
          type: string
        expiryDate:
          type: string
          format: date
      required: [name, category, fileName]
    KnowledgeItem:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        title:
          type: string
        category:
          type: string
          enum: [company_intro, project_description, certification_answer, technical_response, compliance, other]
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        usageCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, category, content]
    CreateKnowledgeItemRequest:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
          enum: [company_intro, project_description, certification_answer, technical_response, compliance, other]
        content:
          type: string
        tags:
          type: array
          items:
            type: string
      required: [title, category, content]
...
    Clarification:
      type: object
      properties:
        id:
          type: integer
        bidId:
          type: integer
        question:
          type: string
        answer:
          type: string
        status:
          type: string
          enum: [pending, submitted, replied]
        submittedAt:
          type: string
          format: date-time
        repliedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, bidId, question, status]
    CreateClarificationRequest:
      type: object
      properties:
        bidId:
          type: integer
        question:
          type: string
      required: [bidId, question]
    UpdateClarificationRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, submitted, replied]
        answer:
          type: string
...