# lib/api-zod/src/generated/types/boqItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Procurement Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BoqItem {
  id: number;
  bidId: number;
  description: string;
  unit?: string;
  quantity: number;
  unitCost: number;
  margin?: number;
  gst?: number;
  totalCost?: number;
  vendorQuote?: number;
  winProbability?: number;
}